feature - Create Transactions, List Accounts, Query transactions in progress

+++ infraestructure/mongodb-repository/src/main/java/bid/dbo/ftracker/repository/data/TransactionData.java
@@ -5,7 +5,6 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -15,6 +14,6 @@ public class TransactionData {
     private String id;
     private String account;
     private Date date;
-    private BigDecimal amount;
+    private Double amount;
     private Transaction.MetaData metaData;
 }
